Create README.md
In this project, you are required to create a simple HTML only website with multiple pages. The website should have following pages:

Homepage
Projects
Articles
Contact
The website should have a navigation bar that should be present on all pages and link to all the pages.

You are not required to style the website, you are only required to create the structure of the website using HTML. Goals of this project are:

Learn how to create multiple pages in a website.
Structure a website using HTML in a semantic way.
Structure in a way that you can easily add styles later.
Add SEO meta tags to the website.